+++ assets/javascripts/kitten/components/graphics/icons-next/icons-next.stories.js
@@ -20,6 +20,8 @@ import {
   UsersIconNext,
   WhatsappIconNext,
   WindowEditIconNext,
+  HeartIconNext,
+  LargeArrowIconNext,
   StatusIconNext,
 } from 'kitten'
 import { Text, Title, pxToRem } from 'kitten'
@@ -139,6 +141,23 @@ export const AllIcons = ({ colorInput }) => {
         </IconContainer>
         <IconContainer>
           <WindowEditIconNext color={colorInput} />
+        </IconContainer>
+        <IconContainer>
+          <HeartIconNext color={colorInput} />
+        </IconContainer>
+      </Group>
+      <Group title="Arrows">
+        <IconContainer suffix='direction="up"'>
+          <LargeArrowIconNext color={colorInput} direction="up" />
+        </IconContainer>
+        <IconContainer suffix='direction="down"'>
+          <LargeArrowIconNext color={colorInput} direction="down" />
+        </IconContainer>
+        <IconContainer suffix='direction="left"'>
+          <LargeArrowIconNext color={colorInput} direction="left" />
+        </IconContainer>
+        <IconContainer suffix='direction="right"'>
+          <LargeArrowIconNext color={colorInput} direction="right" />
         </IconContainer>
       </Group>
       <Group title="Status icons">