+++ assets/javascripts/kitten/components/molecules/comment-block/comment-form/comment-form.stories.js
@@ -1,91 +1,76 @@
 import React from 'react'
 import { CommentForm } from './index'
-import { CheckedCircleIcon } from '../../../..'
 import { action } from '@storybook/addon-actions'
+import { DocsPage } from 'storybook/docs-page'
 
-export const Default = ({ showAvatarBadge, ...args }) => (
-  <CommentForm
-    {...args}
-    avatarBadge={
-      showAvatarBadge && (
-        <span aria-label="Owner" role="aside">
-          <CheckedCircleIcon
-            width="25"
-            height="25"
-            circleColor="#19b4fa"
-            checkedColor="#fff"
-          />
-        </span>
-      )
-    }
-  />
+export default {
+  title: 'Molecules/CommentBlock/CommentForm',
+  component: CommentForm,
+  decorators: [
+    story => (
+      <div className="story-Container story-Grid story-Grid--large">
+        {story()}
+      </div>
+    ),
+  ],
+  parameters: {
+    docs: {
+      page: () => <DocsPage filepath={__filename} importString="CommentForm" />,
+    },
+  },
+  args: {
+    avatarImgProps: {
+      src: `/kitten-${Math.floor(Math.random() * 10)}.jpg`,
+      alt: '',
+    },
+    placeholder: 'Leave a comment',
+    buttonText: 'Comment',
+    error: false,
+    errorMessage: 'Cum sociis natoque penatibus et magnis',
+    onSubmit: action('onSubmit'),
+    onChange: action('onChange'),
+    disabled: false,
+    'aria-label': 'This is the aria label',
+  },
+  argTypes: {
+    avatarImgProps: {
+      name: 'avatarImgProps',
+      control: 'object',
+    },
+    placeholder: {
+      name: 'placeholder',
+      control: 'text',
+    },
+    buttonText: {
+      name: 'buttonText',
+      control: 'text',
+    },
+    error: {
+      name: 'error',
+      control: 'boolean',
+    },
+    errorMessage: {
+      name: 'errorMessage',
+      control: 'text',
+    },
+    disabled: {
+      name: 'disabled',
+      control: 'boolean',
+    },
+    defaultValue: {
+      name: 'defaultValue',
+      control: 'text',
+    },
+    id: {
+      name: 'id',
+      control: 'text',
+    },
+    'aria-label': { control: 'text' },
+  },
+}
+
+export const Default = args => <CommentForm {...args} />
+
+export const WithDefaultValue = args => (
+  <CommentForm {...args} defaultValue="Hello, world" />
 )
-
-Default.decorators = [
-  story => (
-    <div className="story-Container story-Grid story-Grid--large">
-      {story()}
-    </div>
-  ),
-]
-
-Default.args = {
-  avatarImgProps: {
-    src: `/kitten-${Math.floor(Math.random() * 10)}.jpg`,
-    alt: '',
-  },
-  placeholder: 'Leave a comment on the project',
-  commentButton: 'Comment',
-  error: false,
-  errorMessage: 'Cum sociis natoque penatibus et magnis',
-  showAvatarBadge: false,
-  onSubmit: action('onSubmit'),
-}
-
-Default.argTypes = {
-  avatarImgProps: {
-    name: 'avatarImgProps',
-    control: 'object',
-  },
-  placeholder: {
-    name: 'placeholder',
-    control: 'text',
-  },
-  commentButton: {
-    name: 'commentButton',
-    control: 'text',
-  },
-  error: {
-    name: 'error',
-    control: 'boolean',
-  },
-  errorMessage: {
-    name: 'errorMessage',
-    control: 'text',
-  },
-  isDisabled: {
-    name: 'isDisabled',
-    control: 'boolean',
-  },
-  showAvatarBadge: {
-    name: 'showAvatarBadge (story prop)',
-    control: 'boolean',
-  },
-  defaultValue: {
-    name: 'defaultValue',
-    control: 'text',
-  },
-  commentLabel: {
-    name: 'commentLabel',
-    control: 'text',
-  },
-  ariaId: {
-    name: 'ariaId',
-    control: 'text',
-  },
-  avatarBadge: { control: { type: null } },
-  textareaId: {
-    name: 'textareaId',
-    control: 'text',
-  },
-}