+++ assets/javascripts/kitten/components/molecules/comment-block/comment-form/test.js
@@ -1,71 +1,56 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import { CommentForm } from './index'
-import { CheckedCircleIcon } from '../../../..'
-
-const createMockMediaMatcher = matches => () => ({
-  matches,
-  addListener: () => {},
-  removeListener: () => {},
-})
 
 describe('<CommentForm />', () => {
-  let originalMatchMedia
-
-  beforeEach(() => {
-    originalMatchMedia = window.matchMedia
-  })
-
-  afterEach(() => {
-    window.matchMedia = originalMatchMedia
-  })
-
   it('should match its empty snapshot', () => {
-    window.matchMedia = createMockMediaMatcher(false)
-
     const tree = renderer
       .create(
         <CommentForm
-          textareaId="custom-textarea-id"
+          id="custom-textarea-id"
+          aria-label="Leave a comment"
           avatarImgProps={{
             src: '#image',
             alt: 'Image alt',
           }}
           placeholder="Leave a comment on the project"
-          commentButton="Comment"
-          errorMessage="Cum sociis natoque penatibus et magnis"
-          error
+          buttonText="Comment"
         />,
+        {
+          createNodeMock: element => {
+            if (element.type === 'textarea') {
+              return document.createElement('textarea')
+            }
+          },
+        },
       )
       .toJSON()
 
     expect(tree).toMatchSnapshot()
   })
 
-  it('should match a snapshot with a complex avatar badge', () => {
-    window.matchMedia = createMockMediaMatcher(false)
-    const AvatarBadge = () => (
-      <CheckedCircleIcon
-        width="25"
-        height="25"
-        circleColor="#19b4fa"
-        checkedColor="#fff"
-      />
-    )
-
+  it('should match its error snapshot', () => {
     const tree = renderer
       .create(
         <CommentForm
+          id="custom-textarea-id"
+          aria-label="Leave a comment"
           avatarImgProps={{
             src: '#image',
             alt: 'Image alt',
           }}
           placeholder="Leave a comment on the project"
-          commentButton="Comment"
+          buttonText="Comment"
           errorMessage="Cum sociis natoque penatibus et magnis"
           error
-          avatarBadge={<AvatarBadge />}
         />,
+        {
+          createNodeMock: element => {
+            if (element.type === 'textarea') {
+              return document.createElement('textarea')
+            }
+          },
+        },
       )
       .toJSON()