+++ assets/javascripts/kitten/components/molecules/comment-block/comment/index.js
@@ -1,192 +1,250 @@
-import React, { useEffect, useRef, useState } from 'react'
-import styled, { css } from 'styled-components'
+import React from 'react'
+import styled from 'styled-components'
 import PropTypes from 'prop-types'
-import { Marger } from '../../../../components/layout/marger'
-import { CommentAvatar } from '../../../../components/molecules/comment-block/comment-avatar'
+import classNames from 'classnames'
+import deprecated from 'prop-types-extra/lib/deprecated'
+
 import { Text } from '../../../../components/atoms/typography/text'
-import { ScreenConfig } from '../../../../constants/screen-config'
-import COLORS from '../../../../constants/colors-config'
-import { pxToRem, stepToRem } from '../../../../helpers/utils/typography'
-import { LikeButton } from './components/like-button'
-import domElementHelper from '../../../../helpers/dom/element-helper'
+import { mq } from '../../../../constants/screen-config'
+import TYPOGRAPHY from '../../../../constants/typography-config'
+import { pxToRem } from '../../../../helpers/utils/typography'
+import { HeartIconNext } from '../../../../components/graphics/icons-next/heart-icon-next'
 
-const desktopPadding = pxToRem(30)
-const tabletAndMobilePadding = pxToRem(20)
+const CommentWrapper = styled.div`
+  --comment-arrow-size: ${pxToRem(7)};
+  --comment-background-color: var(--color-grey-200);
 
-const StyledContentText = styled(Text)`
-  font-size: ${stepToRem(-1)};
-
-  @media (min-width: ${ScreenConfig.S.min}px) {
-    font-size: ${stepToRem(0)};
-  }
-`
-
-const StyledCommentArrow = styled.span`
-  position: absolute;
-  top: ${pxToRem(20)};
-  display: block;
-  width: 0;
-  height: 0;
-  border-width: ${pxToRem(10)};
-  border-style: solid;
-  border-color: transparent;
-  border-right-color: ${COLORS.background3};
-  left: ${pxToRem(-20)};
-
-  @media (min-width: ${ScreenConfig.S.min}px) {
-    top: ${pxToRem(35)};
-  }
-`
-
-const StyledGrid = styled.div`
   display: flex;
-`
-
-const StyledCommentContainer = styled.span`
+  gap: ${pxToRem(10)};
+  padding: ${pxToRem(15)};
+  border-radius: ${pxToRem(6)};
+  flex-direction: column;
   position: relative;
-  margin-left: ${pxToRem(20)};
-  flex: 1;
+  background-color: var(--comment-background-color);
+  transition: background-color var(--transition);
 
-  @media (min-width: ${ScreenConfig.S.min}px) {
-    margin-left: ${pxToRem(35)};
+  &.k-Comment--isSecondary {
+    margin-left: ${pxToRem(20)};
+
+    @media ${mq.tabletAndDesktop} {
+      margin-left: ${pxToRem(50)};
+    }
   }
-`
+  &.k-Comment--isHighlighted {
+    --comment-background-color: var(--color-grey-400);
 
-const StyledCommentContent = styled.div`
-  border-width: ${pxToRem(2)};
-  background-color: ${COLORS.background3};
-  border-color: ${COLORS.background3};
-  color: ${COLORS.font1};
-  font-size: ${stepToRem(0)};
-  padding: ${tabletAndMobilePadding};
-
-  @media (min-width: ${ScreenConfig.M.min}px) {
-    padding-top: ${desktopPadding};
-    padding-right: ${desktopPadding};
-    padding-left: ${desktopPadding};
+    .k-Comment__header__meta {
+      color: var(--color-grey-700);
+    }
   }
-`
 
-const StyledMargerText = styled.div`
-  ${({ hasLikeButton }) =>
-    hasLikeButton &&
-    css`
-      margin-bottom: ${pxToRem(15)};
+  &::before {
+    content: '';
+    width: 0;
+    height: 0;
+    position: absolute;
+    left: calc(-1 * var(--comment-arrow-size));
+    top: calc(50% - var(--comment-arrow-size));
+    border: var(--comment-arrow-size) solid transparent;
+    border-left: 0;
+    border-right-color: var(--comment-background-color);
+    transition: border-right-color var(--transition);
+  }
 
-      @media (min-width: ${ScreenConfig.M.min}px) {
-        margin-bottom: ${pxToRem(20)};
+  .k-Comment__header {
+    display: flex;
+    gap: ${pxToRem(10)};
+    align-items: center;
+    justify-content: space-between;
+    font-size: ${pxToRem(12)};
+  }
+  .k-Comment__header__meta {
+    display: flex;
+    gap: ${pxToRem(5)};
+    align-items: center;
+    color: var(--color-grey-600);
+  }
+  .k-Comment__header__image {
+    display: block;
+    width: ${pxToRem(20)};
+    height: ${pxToRem(20)};
+    overflow: hidden;
+    object-fit: cover;
+    object-position: center center;
+    border-radius: ${pxToRem(20)};
+  }
+  .k-Comment__header__actions {
+    display: flex;
+    gap: ${pxToRem(10)};
+  }
+  .k-Comment__content {
+    ${TYPOGRAPHY.fontStyles.light}
+    font-size: ${pxToRem(14)};
+  }
+  .k-Comment__footer {
+    display: flex;
+    gap: ${pxToRem(10)};
+  }
+
+  .k-Comment-LikeButton {
+    display: flex;
+    gap: ${pxToRem(5)};
+    align-items: center;
+    color: var(--color-grey-900);
+    transition: color var(--transition);
+
+    :disabled {
+      cursor: not-allowed;
+      opacity: .5;
+    }
+
+    &[aria-pressed] {
+      svg {
+        color: var(--color-red-500);
       }
-    `}
+    }
 
-  ${({ ownerName }) =>
-    ownerName &&
-    css`
-      margin-top: ${pxToRem(10)};
-    `}
-`
+    svg {
+      max-width: ${pxToRem(14)};
+      max-height: ${pxToRem(14)};
+      transition: inherit;
+    }
 
-const StyledBottomNotes = styled(Text)`
-  padding-left: ${pxToRem(30)};
-  margin-top: 0;
-`
-
-const StyledLikeButtonBox = styled.div`
-  position: absolute;
-  right: 0;
-  margin-top: ${pxToRem(-20)};
-  margin-right: ${pxToRem(20)};
-
-  @media (min-width: ${ScreenConfig.M.min}px) {
-    margin-right: ${pxToRem(30)};
+    :not(:disabled):hover {
+      color: var(--color-primary-700);
+    }
   }
 `
 
-export const Comment = ({
-  text,
-  ownerName,
-  avatarImgProps,
-  commentDate,
-  bottomNotes,
-  likeButtonProps,
-  avatarBadge,
-  id,
-}) => {
-  const likeButtonElement = useRef(null)
-  const [likeButtonWidth, setLikeButtonWidth] = useState(0)
+export const Comment = React.forwardRef(
+  (
+    {
+      text,
+      ownerName,
+      ownerUrl,
+      avatarImgProps,
+      commentDate,
+      footer,
+      headerActions,
+      id,
+      children,
+      className,
+      isSecondary,
+      isHighlighted,
+      ...props
+    },
+    ref,
+  ) => (
+    <CommentWrapper
+      ref={ref || null}
+      id={id}
+      className={classNames('k-Comment', className, {
+        'k-Comment--isSecondary': isSecondary,
+        'k-Comment--isHighlighted': isHighlighted,
+      })}
+      {...props}
+    >
+      <div className="k-Comment__header">
+        <div className="k-Comment__header__meta">
+          <a href={ownerUrl}>
+            <img
+              alt=""
+              {...avatarImgProps}
+              className="k-Comment__header__image"
+            />
+          </a>
+          <a href={ownerUrl} className="k-u-link k-u-link-font1 k-u-ellipsis">
+            {ownerName}
+          </a>
+          <Text size="micro" aria-hidden>
+            •
+          </Text>
+          <Text size="micro" weight="light">
+            {commentDate}
+          </Text>
+        </div>
+        <div className="k-Comment__header__actions">{headerActions}</div>
+      </div>
 
-  useEffect(() => {
-    const marginLeftAndRight = 40
-    const elementWidth = domElementHelper.getComputedWidth(
-      likeButtonElement.current,
-    )
-
-    setLikeButtonWidth(elementWidth + marginLeftAndRight)
-  }, [])
-
-  return (
-    <StyledGrid id={id}>
-      <CommentAvatar
-        avatarImgProps={avatarImgProps}
-        commentDate={commentDate}
-        avatarBadge={avatarBadge}
-      />
-      <StyledCommentContainer>
-        <StyledCommentContent>
-          {ownerName && (
-            <Marger bottom="1">
-              <Text color="font1" size="tiny" weight="regular">
-                {ownerName}
-              </Text>
-            </Marger>
-          )}
-
-          <StyledMargerText
-            hasLikeButton={!!likeButtonProps.children}
-            ownerName={ownerName}
-          >
-            <StyledContentText color="font1" weight="light">
-              {text}
-            </StyledContentText>
-          </StyledMargerText>
-          <StyledCommentArrow />
-        </StyledCommentContent>
-
-        {likeButtonProps.children && (
-          <StyledLikeButtonBox ref={likeButtonElement}>
-            <LikeButton {...likeButtonProps} />
-          </StyledLikeButtonBox>
+      <div className="k-Comment__content">
+        {!!text && (
+          <Text color="font1" weight="light" size="tiny">
+            {text}
+          </Text>
         )}
 
-        {bottomNotes && (
-          <Marger top=".5" style={{ marginRight: likeButtonWidth }}>
-            <StyledBottomNotes
-              tag="div"
-              color="font1"
-              size="micro"
-              weight="bold"
-            >
-              {bottomNotes}
-            </StyledBottomNotes>
-          </Marger>
-        )}
-      </StyledCommentContainer>
-    </StyledGrid>
-  )
-}
+        {children}
+      </div>
+
+      {footer && <div className="k-Comment__footer">{footer}</div>}
+    </CommentWrapper>
+  ),
+)
+
+Comment.LikeButton = ({
+  children,
+  accessibilityLabel,
+  hasLiked,
+  className,
+  disabled,
+  ...props
+}) => (
+  <button
+    role="button"
+    aria-pressed={!!hasLiked || null}
+    className={classNames(
+      'k-Comment-LikeButton',
+      'k-u-reset-button',
+      'k-u-link',
+      className,
+    )}
+    disabled={disabled}
+    {...props}
+  >
+    <span className="k-u-a11y-visuallyHidden">{accessibilityLabel}</span>
+    {children}
+    <HeartIconNext
+      aria-hidden="true"
+      focusable="false"
+      color="currentColor"
+      bgColor={hasLiked ? 'currentColor' : 'transparent'}
+    />
+  </button>
+)
 
 Comment.propTypes = {
-  text: PropTypes.node.isRequired,
   ownerName: PropTypes.string,
-  avatarImgProps: PropTypes.object.isRequired,
-  commentDate: PropTypes.string.isRequired,
-  bottomNotes: PropTypes.node,
-  likeButtonProps: PropTypes.object,
-  avatarBadge: PropTypes.node,
+  ownerUrl: PropTypes.string,
+  avatarImgProps: PropTypes.object,
+  commentDate: PropTypes.string,
+  headerActions: PropTypes.node,
+  footer: PropTypes.node,
+  bottomNotes: deprecated(PropTypes.node, 'Please use footer prop instead'),
+  text: deprecated(PropTypes.node, 'Please use children prop instead'),
+  likeButtonProps: deprecated(
+    PropTypes.object,
+    'Please use headerActions={<Comment.LikeButton />} instead',
+  ),
+  isSecondary: PropTypes.bool,
+  isHighlighted: PropTypes.bool,
 }
 
 Comment.defaultProps = {
-  bottomNotes: '',
+  footer: null,
   ownerName: '',
-  likeButtonProps: {},
-  avatarBadge: '',
+  ownerUrl: '',
+  headerActions: null,
+  avatarImgProps: {},
+  commentDate: '',
+  isSecondary: false,
+  isHighlighted: false,
+}
+
+Comment.LikeButton.defaultProps = {
+  hasLiked: false,
+}
+
+Comment.LikeButton.propTypes = {
+  accessibilityLabel: PropTypes.string.isRequired,
+  hasLiked: PropTypes.bool,
 }