+++ assets/javascripts/kitten/components/molecules/comment-block/comment/test.js
@@ -1,28 +1,9 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import { Comment } from './index'
-import { CheckedCircleIcon } from '../../../..'
-
-const createMockMediaMatcher = matches => () => ({
-  matches,
-  addListener: () => {},
-  removeListener: () => {},
-})
 
 describe('<Comment />', () => {
-  let originalMatchMedia
-
-  beforeEach(() => {
-    originalMatchMedia = window.matchMedia
-  })
-
-  afterEach(() => {
-    window.matchMedia = originalMatchMedia
-  })
-
   it('should match its empty snapshot', () => {
-    window.matchMedia = createMockMediaMatcher(false)
-
     const tree = renderer
       .create(
         <Comment
@@ -31,13 +12,18 @@ describe('<Comment />', () => {
             alt: 'Image alt',
           }}
           commentDate="Custom date"
-          text="Custom text"
           ownerName="Custom name"
-          likeButtonProps={{
-            children: '4',
-            hasLiked: true,
-          }}
-        />,
+          headerActions={
+            <Comment.LikeButton
+              hasLiked
+              accessibilityLabel="Retirer des favoris"
+            >
+              4
+            </Comment.LikeButton>
+          }
+        >
+          Custom text
+        </Comment>,
       )
       .toJSON()
 
@@ -45,7 +31,6 @@ describe('<Comment />', () => {
   })
 
   it('should match a snapshot with a complex bottom notes', () => {
-    window.matchMedia = createMockMediaMatcher(false)
     const BottomNotes = () => (
       <div>
         <a href="#" style={{ color: '#000' }}>
@@ -65,10 +50,11 @@ describe('<Comment />', () => {
             alt: 'Image alt',
           }}
           commentDate="Custom date"
-          text="Custom text"
           ownerName="Custom name"
-          bottomNotes={<BottomNotes />}
-        />,
+          footer={<BottomNotes />}
+        >
+          Custom text
+        </Comment>,
       )
       .toJSON()
 
@@ -76,16 +62,6 @@ describe('<Comment />', () => {
   })
 
   it('should match a snapshot with a complex avatar badge', () => {
-    window.matchMedia = createMockMediaMatcher(false)
-    const AvatarBadge = () => (
-      <CheckedCircleIcon
-        width="25"
-        height="25"
-        circleColor="#19b4fa"
-        checkedColor="#fff"
-      />
-    )
-
     const tree = renderer
       .create(
         <Comment
@@ -94,10 +70,10 @@ describe('<Comment />', () => {
             alt: 'Image alt',
           }}
           commentDate="Custom date"
-          text="Custom text"
           ownerName="Custom name"
-          avatarBadge={<AvatarBadge />}
-        />,
+        >
+          Custom text
+        </Comment>,
       )
       .toJSON()