+++ CHANGELOG.md
@@ -4,6 +4,32 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Tokens: Add `transitions` related tokens.
+- Feature: `Comment`:
+  - Change design
+  - Deprecate the following props:
+    - `bottomNotes` replaced by `footer`
+    - `text` replaced by `children`
+    - `likeButtonProps` replaced by `headerActions={<Comment.LikeButton/>}`
+    - `avatarBadge`
+- Feature: `CommentForm`:
+  - Modernize component code
+  - Change design
+  - Deprecate the following props:
+    - `commentButton` replaced by `buttonText`
+    - `isDisabled` replaced by `deleted`
+    - `commentLabel` replaced by `aria-label`
+    - `ariaId` replaced by `id`
+    - `textareaId` replaced by `id`
+    - `avatarBadge`
+- Feature: `ModalFooterInput`: Add component.
+- Feature: `HeartIconNext`: Add component.
+- Feature: `LargeArrowIconNext`: Add component.
+- Fix: Remove console errors on `DownloadIcon` and `LinkedinIconNext`.
+- Feature: Utilities: Add `k-u-ellipsis` CSS utility class.
+- Feature: Utilities: Add `baseline` option to `k-u-alignItems-`  CSS utility classes.
+- Feature: Utilities: Add `grey1` option to `k-u-link-` CSS utility classes.
+- Feature: `TextareaAutoResize`: Add component.
 - Fix: `TagInput`: Fix focus styles.
 
 ## [6.5.1] - 2021-11-30